ioctls: move VfioError into lib.rs
Move VfioError into lib.rs, and a unit test case.
